ファイルを分けてコマンドを実行できるようにしました
